Make `Depth` store an `i32` instead of an `i16`.
Overflows can sometimes happen otherwise, as multiplying two `u8`s
sometimes doesn't fit in an `i16`.
No speed difference.

Bench: 12386906

+++ build.rs
@@ -25,7 +25,7 @@ use std::{
 };
 
 /// The difference between the leaf node and the current node.
-type Depth = i16;
+type Depth = i32;
 /// A table of base late move reductions.
 ///
 /// Indexed by the depth then number of legal moves.
